backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
shapely==2.0.6
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pydantic==2.8.2
huggingface_hub==0.23.5
